feat: add container prefix, add createnetwork to traefik
